+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - `#[use_modtype]` attribute macro.
 - Trait specific attributes.
 - `modtype::Pow` for `{u64, u128, i8, i16, i32, i64, i128, isize}`
+- `modtype::new_unchecked`.
 
 ### Changed